impr: emerald Liquid Glass on iOS 26 and macOS 26
Color Safari's UI on iOS and macOS 26 just like on iOS 18 and macOS
Sequoia.

Setting the background color of the `<html>` or `<body>` element will
change the color of the window chrome on both current and new OSes. The
`<meta name="theme-color">` will change the color of the status bar on
iOS and change the color of the tab bar on macOS, but it doesn't have
any effect on the new OSes. On the iOS 18, you can set the background
color of the `<html>` or `<body>` elements and it will color the tab bar
separately.

Since we're using the same color for both "theme-color" and the
`<body>`, the `<meta name="theme-color">` element is kinda redundant but
we're gonna keep it anyway...