feat: hacked a drop-shadow under the header
So adding a drop shadow on the `<header>` element doesn't work because
it has the masks applied to it. So I hardcoded the width and the height
of a div, put it under the header, and blurred the hell out of it since
I know it works.

There's probably a better way to to do this...

At least we should put the blurred div in the Header component itself
since it makes more sense for it to be there...