chore(lib): vendored source code
This code is from my personal site. They are used to write JSX and turn
it into SVGs and PNGs using Satori and resvg.

I'm in the middle of working on a PR for a minimal JSX runtime for
Satori but I still want to try it out in other places to make sure that
its solid before marking it ready for review.