fix(css): use `clip-path` for header
Create a new 'lib' module that does shows the how the coordinates are
calculated for the polygon shape.

The `mask-image` thing that I was doing with gradients was pretty hacky
and I was getting some subtle rendering issues on my phone running iOS
26, but this solves that. Much better browser support using `clip-path`
and `polygon()` anyway.

The clip path also adheres to the dimensions outlined in SN0101 so the
header now has a more accurate street sign shape than before where I was
just eye-balling it and doing some loosey-goosey math.