refactor(scripts): changes to `calver-utils`
Renamed functions, changed some terminology, and updated CalVer scheme
RegEx.

According to the SemVer spec, "v1.2.3" is not a semantic version but
"1.2.3" is. The "v" prefix is a common way (in English) to indicate a
version number and commonly used in version control. So in keeping with
this convention, refactor the `calver-utils` module to not handle the
"v" prefix for Git tag names and just the version identifier itself.

https://semver.org/